Fix minor bug showing connection rates
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@3354 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/au/network/Service.cpp
@@ -158,10 +158,10 @@ namespace au
                 values.push_back( item->socket_connection->getHostAndPort() );
                 values.push_back( au::str("%lu", item->socket_connection->getTime() ) );
                 
-                values.push_back( au::str("%lu", item->socket_connection->rate_in.getTotalSize() ) );
-                values.push_back( au::str("%lu", item->socket_connection->rate_out.getTotalSize() ) );
-                values.push_back( au::str("%lu", item->socket_connection->rate_in.getRate() ) );
-                values.push_back( au::str("%lu", item->socket_connection->rate_out.getRate() ) );
+                values.push_back( au::str("%lu", (size_t)item->socket_connection->rate_in.getTotalSize() ) );
+                values.push_back( au::str("%lu", (size_t)item->socket_connection->rate_out.getTotalSize() ) );
+                values.push_back( au::str("%lu", (size_t)item->socket_connection->rate_in.getRate() ) );
+                values.push_back( au::str("%lu", (size_t)item->socket_connection->rate_out.getRate() ) );
                 
                 table->addRow( values );
             }
